assignment 2 done

+++ src/modules/vehicles/vehicles.service.ts
@@ -83,6 +83,16 @@ const updateVechicles = async (
 };
 
 const deleteVehicles = async (id: string) => {
+    const activeBooking = await pool.query(
+        `SELECT id FROM bookings WHERE vehicle_id  = $1 AND status = 'active'`,
+        [id]
+    );
+
+    if (activeBooking.rows.length > 0) {
+        throw new Error("Vehicle cannot be deleted, active booking exist");
+    }
+
+    // delete vehicle
     const result = await pool.query(`DELETE FROM vehicles WHERE id = $1`, [id]);
 
     return result;
